Fix live activity UNION query and remove empty journey placeholder panel.
Wrap each UNION branch in parentheses for valid Postgres syntax; only show the journey panel after a lead is selected.

+++ src/lib/recent-activity.ts
@@ -27,81 +27,81 @@ export async function fetchRecentActivity(limit = 40): Promise<RecentActivityIte
   const { rows } = await pool.query(
     `
     WITH recent AS (
-      SELECT 'sale' AS action_type,
-             'hyros_sale:' || id::text AS event_id,
-             lead_email AS email,
-             COALESCE(product_name, 'Sale') AS title,
-             created_at,
-             'hyros' AS source_system,
-             product_price::float8 AS value
-      FROM fct_hyros_attributed_sales
-      ORDER BY created_at DESC
-      LIMIT $1
+      (SELECT 'sale' AS action_type,
+              'hyros_sale:' || id::text AS event_id,
+              lead_email AS email,
+              COALESCE(product_name, 'Sale') AS title,
+              created_at,
+              'hyros' AS source_system,
+              product_price::float8 AS value
+       FROM fct_hyros_attributed_sales
+       ORDER BY created_at DESC
+       LIMIT $1)
 
       UNION ALL
 
-      SELECT 'optin',
-             'hyros_optin:' || id::text,
-             lead_email,
-             COALESCE(event_type, 'Opt-in'),
-             created_at,
-             'hyros',
-             NULL
-      FROM hyros_lead_opt_ins
-      ORDER BY created_at DESC
-      LIMIT $1
+      (SELECT 'optin',
+              'hyros_optin:' || id::text,
+              lead_email,
+              COALESCE(event_type, 'Opt-in'),
+              created_at,
+              'hyros',
+              NULL::float8
+       FROM hyros_lead_opt_ins
+       ORDER BY created_at DESC
+       LIMIT $1)
 
       UNION ALL
 
-      SELECT 'call',
-             'hyros_call:' || id::text,
-             lead_email,
-             COALESCE(event_type, 'Call'),
-             created_at,
-             'hyros',
-             NULL
-      FROM hyros_attributed_calls
-      ORDER BY created_at DESC
-      LIMIT $1
+      (SELECT 'call',
+              'hyros_call:' || id::text,
+              lead_email,
+              COALESCE(event_type, 'Call'),
+              created_at,
+              'hyros',
+              NULL::float8
+       FROM hyros_attributed_calls
+       ORDER BY created_at DESC
+       LIMIT $1)
 
       UNION ALL
 
-      SELECT 'membership',
-             'whop:' || id::text,
-             email,
-             COALESCE(product_title, plan_id, 'Membership'),
-             COALESCE(whop_created_at, created_at),
-             'whop',
-             NULL
-      FROM whop_memberships
-      ORDER BY COALESCE(whop_created_at, created_at) DESC
-      LIMIT $1
+      (SELECT 'membership',
+              'whop:' || id::text,
+              email,
+              COALESCE(product_title, plan_id, 'Membership'),
+              COALESCE(whop_created_at, created_at),
+              'whop',
+              NULL::float8
+       FROM whop_memberships
+       ORDER BY COALESCE(whop_created_at, created_at) DESC
+       LIMIT $1)
 
       UNION ALL
 
-      SELECT 'triage',
-             'triage:' || id::text,
-             lead_email,
-             COALESCE(source, 'Booking') || ' (' || COALESCE(status, 'open') || ')',
-             COALESCE(start_time, created_at),
-             'triage',
-             NULL
-      FROM triage_tickets
-      ORDER BY COALESCE(start_time, created_at) DESC
-      LIMIT $1
+      (SELECT 'triage',
+              'triage:' || id::text,
+              lead_email,
+              COALESCE(source, 'Booking') || ' (' || COALESCE(status, 'open') || ')',
+              COALESCE(start_time, created_at),
+              'triage',
+              NULL::float8
+       FROM triage_tickets
+       ORDER BY COALESCE(start_time, created_at) DESC
+       LIMIT $1)
 
       UNION ALL
 
-      SELECT 'lead',
-             'lead:' || email,
-             email,
-             'Lead created',
-             created_at,
-             'crm',
-             NULL
-      FROM leads
-      ORDER BY created_at DESC
-      LIMIT $1
+      (SELECT 'lead',
+              'lead:' || email,
+              email,
+              'Lead created',
+              created_at,
+              'crm',
+              NULL::float8
+       FROM leads
+       ORDER BY created_at DESC
+       LIMIT $1)
     )
     SELECT r.event_id AS id,
            r.email,
